backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg[binary]==3.1.18
langchain==0.2.3
openai==1.30.1
python-dotenv==1.0.1
pgvector
sqlalchemy[asyncio]==2.0.19
asyncpg==0.28.0

# Development dependencies
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.1
ruff==0.0.280
black==23.7.0
pre-commit==3.3.3 